Factor out generateNodeBuilderFunctions to build-utils
